+++ src/templates/javascript/reactArrowComponent.ts
@@ -1,13 +1,16 @@
 import CreateComponent from '../interfaces/CreateComponent';
 import creatReactImport from '../shared/functions/create-react-import';
+import pascalCase from '../shared/functions/pascal-case';
 
-export default ({ componentName, useReactImport }: CreateComponent) => (
-`${ creatReactImport(useReactImport) }const ${ componentName } = () => {
+export default ({ componentName, useReactImport, useExportDefault }: CreateComponent) => (
+`${ creatReactImport(useReactImport) }${useExportDefault ? '' : 'export '}const ${ pascalCase(componentName) } = () => {
   return (
-    <h1>${ componentName }</h1>
+    <h1>${ pascalCase(componentName) }</h1>
   );
 }
+${useExportDefault ? `
 
-export default ${ componentName };
+export default ${ pascalCase(componentName) };
+` : ''}
 `
 );