+++ src/templates/javascript/reactComponent.ts
@@ -2,13 +2,15 @@ import CreateComponent from '../interfaces/CreateComponent';
 import creatReactImport from '../shared/functions/create-react-import';
 import pascalCase from '../shared/functions/pascal-case';
 
-export default ({ componentName, useReactImport }: CreateComponent) => (
-`${ creatReactImport(useReactImport) }function ${ pascalCase(componentName) }() {
+export default ({ componentName, useReactImport, useExportDefault }: CreateComponent) => (
+`${ creatReactImport(useReactImport) }${useExportDefault ? '' : 'export '}function ${ pascalCase(componentName) }() {
   return (
     <h1>${ pascalCase(componentName) }</h1>
   );
 }
+${useExportDefault ? `
 
 export default ${ pascalCase(componentName) };
+` : ''}
 `
 );