+++ src/templates/shared/functions/barrel.ts
@@ -1,0 +1,7 @@
+import CreateComponent from "../../interfaces/CreateComponent";
+import pascalCase from "./pascal-case";
+
+export default ({ componentName }: CreateComponent) =>
+  `
+export * from './${componentName}';
+`;