+++ src/templates/typescript/reactComponent.ts
@@ -2,12 +2,12 @@ import CreateComponent from '../interfaces/CreateComponent';
 import creatReactImport from '../shared/functions/create-react-import';
 import pascalCase from '../shared/functions/pascal-case';
 
-export default ({ componentName, useReactImport }: CreateComponent) => (
+export default ({ componentName, useReactImport, useExportDefault }: CreateComponent) => (
 `${ creatReactImport(useReactImport, true) }interface ${ pascalCase(componentName) }Props {
   children: ReactNode;
 }
 
-function ${ pascalCase(componentName) }({ children }: ${ pascalCase(componentName) }Props) {
+${useExportDefault ? '' : 'export '}function ${ pascalCase(componentName) }({ children }: ${ pascalCase(componentName) }Props) {
   return (
     <>
       <h1>${ pascalCase(componentName) }</h1>
@@ -15,7 +15,9 @@ function ${ pascalCase(componentName) }({ children }: ${ pascalCase(componentNam
     </>
   );
 }
+${useExportDefault ? `
 
 export default ${ pascalCase(componentName) };
+` : ''}
 `
 );