+++ src/templates/typescript/reactNativeComponent.ts
@@ -2,14 +2,14 @@ import CreateComponent from '../interfaces/CreateComponent';
 import creatReactImport from '../shared/functions/create-react-import';
 import pascalCase from '../shared/functions/pascal-case';
 
-export default ({ componentName, useReactImport }: CreateComponent) => (
+export default ({ componentName, useReactImport, useExportDefault }: CreateComponent) => (
 `${ creatReactImport(useReactImport, true) }import { View, Text, StyleSheet } from 'react-native';
 
 interface ${ pascalCase(componentName) }Props {
   children: ReactNode;
 }
 
-function ${ pascalCase(componentName) }({ children }: ${ pascalCase(componentName) }Props) {
+${useExportDefault ? '' : 'export '}function ${ pascalCase(componentName) }({ children }: ${ pascalCase(componentName) }Props) {
   return (
     <View style={ styles.container }>
       <Text style={ styles.title }>${ pascalCase(componentName) }</Text>
@@ -31,7 +31,9 @@ const styles = StyleSheet.create({
     color: '#fff',
   },
 });
+${useExportDefault ? `
 
 export default ${ pascalCase(componentName) };
+` : ''}
 `
 );