+++ src/templates/typescript/styledReactNativeComponent.ts
@@ -2,7 +2,7 @@ import CreateComponent from '../interfaces/CreateComponent';
 import creatReactImport from '../shared/functions/create-react-import';
 import pascalCase from '../shared/functions/pascal-case';
 
-export default ({ componentName, useReactImport }: CreateComponent) => (
+export default ({ componentName, useReactImport, useExportDefault }: CreateComponent) => (
 `${ creatReactImport(useReactImport, true) }import { Text } from 'react-native';
 
 import { Container } from './styles';
@@ -11,15 +11,17 @@ interface ${ pascalCase(componentName) }Props {
   children: ReactNode;
 }
 
-function ${ pascalCase(componentName) }({ children }: ${ pascalCase(componentName) }Props) {
+${useExportDefault ? '' : 'export '}function ${ pascalCase(componentName) }({ children }: ${ pascalCase(componentName) }Props) {
   return (
     <Container>
-      <Text>${ componentName }</Text>
+      <Text>${ pascalCase(componentName) }</Text>
       {children}
     </Container>
   );
 };
+${useExportDefault ? `
 
-export default ${ componentName };
+export default ${ pascalCase(componentName) };
+` : ''}
 `
 );